PowerMeter.py: improve Error messages
Stripped the command variable to not display any control
character (i.e. newline) to users.

Removed the transport object name as it is of little value
to users.

Improved message's readability by introducing a newline character
that splits it.

Before we would get a message like:
  $ ./yokotool /dev/usbtmc0 read V -l 2
  yokotool: ERROR: failed to write command "
  *CLS" to <yokotools.transport.USBTMC object at 0x7f809de34190>: error while writing to device '/dev/usbtmc0': [Errno 9] Bad file descriptor

And now we get:
  $ ./yokotool /dev/usbtmc0 read V -l 2
  yokotool: ERROR: failed to write command "*CLS" to the power meter:
  error while writing to device '/dev/usbtmc0': [Errno 9] Bad file descriptor

Change-Id: I518be4b8bb31cb3ee3290c23b84a9a71fbe8580e
Signed-off-by: Helia Correia <helia.correia@intel.com>