yokotool: add SMOOTHING_PROPERTIES data structure
Added data structure that describes smoothing properties
and maps to the related power meter's command.

---

This commit is the first of a series of patches that aim
at improving the organization of smoothing's related commands.

-----------
BIG PICTURE
-----------

WT310 power meters support 3 smoothing commands:
  - smoothing (or smoothing status)
  - smoothing type
  - smoothing factor

Our approach consists into apprehending each of these commands
like a smoothing property (or subcommand), and group them all
under the "smoothing" command.

The syntax to act on one of these properties will be as follows:
  - Get the value of a property:
    $ yokotool <device_node> smoothing <status|type|factor>
  - Set the value of a property:
    $ yokotool <device_node> smoothing <status|type|factor> --value <value>

----
NOTE
----

This improvement is split into several commits to ease the review.
Each commit may not function properly if tested separately.

Change-Id: I85adf16c41d2f2456ba5d06fb8b59bfee5bb180d
Signed-off-by: Helia Correia <helia.correia@intel.com>