yokotool: add "smoothing" command to CMDLINE_COMMANDS data structure
Added a new smoothing command to the list of existing commands
along with the adequate help and description messages.

---

This commit is part of a series of patches that aim at
improving the organization of smoothing's related commands.

-----------
BIG PICTURE
-----------

WT310 power meters support 3 smoothing commands:
  - smoothing (or smoothing status)
  - smoothing type
  - smoothing factor

Our approach consists into apprehending each of these commands
like a smoothing property (or subcommand), and group them all
under the "smoothing" command.

The syntax to act on one of these properties will be as follows:
  - Get the value of a property:
    $ yokotool <device_node> smoothing <status|type|factor>
  - Set the value of a property:
    $ yokotool <device_node> smoothing <status|type|factor> --value <value>

----
NOTE
----

This improvement is split into several commits to ease the review.
Each commit may not function properly if tested separately.

Change-Id: Id978007c670f7835a37a48918879640119135fea
Signed-off-by: Helia Correia <helia.correia@intel.com>