yokotool: support "?" option for "smoothing" command
"?" option returns the list of properties to get or set for
"smoothing" command.

---

This commit is part of a series of patches that aim at
improving the organization of smoothing's related commands.

-----------
BIG PICTURE
-----------

WT310 power meters support 3 smoothing commands:
  - smoothing (or smoothing status)
  - smoothing type
  - smoothing factor

Our approach consists into apprehending each of these commands
like a smoothing property (or subcommand), and group them all
under the "smoothing" command.

The syntax to act on one of these properties will be as follows:
  - Get the value of a property:
    $ yokotool <device_node> smoothing <status|type|factor>
  - Set the value of a property:
    $ yokotool <device_node> smoothing <status|type|factor> --value <value>

----
NOTE
----

This improvement is split into several commits to ease the review.
Each commit may not function properly if tested separately.

Change-Id: I6b2210085b78af220a5899d1c1e84e22f40dcde3
Signed-off-by: Helia Correia <helia.correia@intel.com>