yokotool: switch "smoothing --value" to an optional positional parameter
"--value" optional parameter seems to be useless and there is no point
complicating the command line interface with long options.

So switching to using "value" optional positional parameter.

And getting rid of the "str()" typecasting of "args.value" which is not
required anymore.

---

This commit is part of a series of patches that aim at
improving the organization of smoothing's related commands.

-----------
BIG PICTURE
-----------

WT310 power meters support 3 smoothing commands:
- smoothing (or smoothing status)
- smoothing type
- smoothing factor

Our approach consists into apprehending each of these commands
like a smoothing property (or subcommand), and group them all
under the "smoothing" command.

The syntax to act on one of these properties is now as follows:
  - Get the value of a property:
    $ yokotool <device_node> smoothing <status|type|factor>
  - Set the value of a property:
    $ yokotool <device_node> smoothing <status|type|factor> <value>

Change-Id: I1e2bd5eeed077d963fbb37b825e360d086a26966
Signed-off-by: Helia Correia <helia.correia@intel.com>