yokotool: support "?" option for "smoothing" subcommands
"?" option returns the list of values to get or set for
"smoothing status", "smoothing type" and "smoothing factor"
subcommands.

---

This commit is part of a series of patches that aim at
improving the organization of smoothing's related commands.

-----------
BIG PICTURE
-----------

WT310 power meters support 3 smoothing commands:
  - smoothing (or smoothing status)
  - smoothing type
  - smoothing factor

Our approach consists into apprehending each of these commands
like a smoothing property (or subcommand), and group them all
under the "smoothing" command.

The syntax to act on one of these properties is now as follows:
  - Get the value of a property:
    $ yokotool <device_node> smoothing <status|type|factor>
  - Set the value of a property:
    $ yokotool <device_node> smoothing <status|type|factor> <value>

Change-Id: Iabd820437a694203d02fcdfe1aee204585465d9f
Signed-off-by: Helia Correia <helia.correia@intel.com>