Support a configuration file
Support a "/home/$USER/.yokotool.conf" file containing the device model
along with the device node.

A configuration file spares users the need to specify these information
within the command-line. However, when they do so, these information
prevail over the corresponding data from the configuration file.

The configuration file format is "ini" as its syntax is easy to read
and write, and the file format is supported by the stdlib (no need
to install any extra package).

Also updated the man page, TODO file, and added a template configuration
file to be used as a basis.

Change-Id: Iedbcfef67b02b1ab16412d53dfe8891f786955e0
Signed-off-by: Helia Correia <helia.correia@linux.intel.com>

Conflicts:
	TODO