yokotool: improve the handling of "--list-models" argument
Improved the docstrings describing ListModelsAction class and __call__()
function.

Removed the call to setattr() function since it's useless - the function
resets the attribute of an object to a value that is already assigned.

Fixed a bug - now we list all the supported power meters models and
not just the first element of the list.

Change-Id: If0906bdf35bf93634666bf14b05bdaf974e61cc1
Signed-off-by: Helia Correia <helia.correia@linux.intel.com>