Revert the support of "--list-model" command-line option
Reverted the support of "--list-model" command-line option because
as long as only one power meter is supported (WT310), this option
makes little sense and can only bring confusion.

The implementation of "--list-model" is part of the support of WT210
power meters model which is not mature enough to go out yet.

Change-Id: I9a3846fa8eaa042d03e4d71e885be70ce6cf389c
Signed-off-by: Helia Correia <helia.correia@linux.intel.com>