yokotool: set a default value for -m | --devmodel
It's generally a bad approach to break what worked
previously and to make users re-learn how to use a
software.

The introduction of -m | --devmodel had such an
unfortunate impact on Yoko tool.

Since yokotool only supports one model of power meter,
it should be smart enough to know which one it is and
not oblige users to mention it.

Also removed the str() typecast which is now useless.

Change-Id: I880ff91f60b08bcb785561feda22c1a04ec10a65
Signed-off-by: Helia Correia <helia.correia@linux.intel.com>