yokotool: BUG FIX: enable users to set an integration timer value
get_argument_help() usually returns the valid values for a property.

So we used to rely on this function before setting an integration
property.

However, integration timer was an exception to that rule because
get_argument_help() would return a message for it, and not a range
of valid values for instance.

With this patch we verify the timer value against a range of
integers instead of a help message.

Change-Id: I45525be75630922b26995962e6949aeb635827d2
Signed-off-by: Helia Correia <helia.correia@linux.intel.com>