(p2to3) transport.py: encode/decode ASCII strings into bytes
"Python 3.0 uses the concepts of text and (binary) data instead
of Unicode strings and 8-bit strings."

So we need to turn strings into bytes when writing data, and
bytes into strings when reading data.

Source https://docs.python.org/3.0/whatsnew/3.0.html

====
NOTE
====

This patch is part of the series (tagged with "p2to3") that
aim at porting yoko-tool to Python3.

Change-Id: I384a5bff059691a21736fe899b9b14b38cd6da7c
Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>