yokotool: BUG FIX: use "error_stream" instead of "sys.stderr"
In setup_logger() we want to color the error messages that we
write to a terminal. The method provides "error_stream"
parameter that can be used to define a custom stream for error
messages. So we should check if this parameter is an interactive
stream or not, instead of checking "sys.stderr".

To properly fix this bug, we added 2 attributes to our log
object, namely "output_stream" and "error_stream". And we use
them where applicable.

Change-Id: I3f13c676d6013c7333b105e6802e8f76b1eb1f76
Signed-off-by: Helia Correia <helia.correia@linux.intel.com>