(p2to3) yokotool: fix "listen" command breakages
In Python3, the default mode when calling makefile()
is "r" so we set it to "w" since we both read and
write to the socket.

Source https://docs.python.org/3/library/socket.html#socket.socket.makefile

Also, since we read bytes from the socket, we now need to
decode this stream as a "utf-8" string.

Change-Id: Iad69a8535c447f8da6ba060448586e14a3881f7d
Signed-off-by: Helia Correia <helia.correia@linux.intel.com>