Rename "PowerMeter.py" into "YokoBase.py"
Up to now, PowerMeter class was extended by WT310 class and the
classes hierarchy was like:
  PowerMeter -> WT310

Now we're planning to change the architecture of Yoko Tool's
libraries and introduce a new module named "PowerMeter.py".

So the classes hierarchy will be like:
  YokoBase (old PowerMeter) -> WT310 -> PowerMeter (new module)

---

This patch is part of the series that prepare/implement
a way to provide virtual data (like time stamp, Joules)
along with the data we read from the power meter (like
power, current, voltage among others).

Change-Id: I46018c62b2cf6f27bd7f3b7376e85cad3bec9481
Signed-off-by: Helia Correia <helia.correia@linux.intel.com>