Move "set_data_items()" to PowerMeter.py
The setting of data items shouldn't belong to yokotool
because it's a command-line interface.

The complexity induced by the handling of commands
should hide behind the curtains.

---

This patch is part of the series that prepare/implement
a way to provide virtual data (like time stamp, Joules)
along with the data we read from the power meter (like
power, current, voltage among others).

Signed-off-by: Helia Correia <helia.correia@linux.intel.com>

Change-Id: I9cfa3538df95b4af74ffa84a819c1a0e9c8a13e0