PowerMeter.py: get the data items to read
Get the list of data items to be read from the power
meter. If users request to read the 'Joules' virtual
data item, we read 'Power' and will use this value
to later compute the Joules.

---

This patch is part of the series that prepare/implement
a way to provide virtual data (like time stamp, Joules)
along with the data we read from the power meter (like
power, current, voltage among others).

Change-Id: Ia9a3b8cd20dafd98f245d6df684634621fccae07
Signed-off-by: Helia Correia <helia.correia@linux.intel.com>