Add imperfect WT320 and WT330 support
WT332 and WT333 are same as WT310, but have 2 or 3 elements. Ideally we need to
support all elements, but with this small patch we may support these
powermeters in a limited mode - only one measurement element is supported..

May be in the future we can support all the 3 elements, this would need more
work.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>