yokolibs: use word 'parse' instead of 'process'
This is a cosmetic commit that renames few functions in '_config.py' so that
hey use word 'parse' instead of 'process'. "Process the configuration file" is
OK, but sounds a bit like there is a lot of work to do, while all we do is just
parsing it. "Parse" seems a bit more suitable.

Change-Id: I5a0267945585b6d857043fcaab70f1c20b212769
Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>