Logging: add logging to files
Add a capability of logging to both the console and log files.
We do not need this functionality in yoko-tools, but the 'Logging' module that
we use here is the same as what we use in other projects (e.g., 'wult'), so we
keep them in sync. So basically this patch is about syncing the 2 copies of the
same code between 2 projects.

Change-Id: Id81879933e47e7b0952b1908a298693329de6f19
Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>